+++ check_save.php
@@ -12,6 +12,17 @@ if($_POST['mode']=="add"){
     mysqli_query($connect, $sql);
     echo "<script>alert('แก้ไขข้อมูลเรียบร้อย');</script>";
 }elseif($_GET['mode']=="delete"){
+
+    $sql_id1 = "select * from repair_checklist where ch_id='".$_GET['id']."'";
+    $query_id1 = mysqli_query($connect, $sql_id1);
+    $num1 = mysqli_num_rows($query_id1);
+
+    if($num1 >=1){
+        echo "<script>alert('ไม่สามารถลบข้อมูลได้ เนื่องจากข้อมูลนี่ถูกใช้แล้ว');</script>";
+        echo "<META http-equiv='refresh' Content='0; URL=check_list.php'> ";
+        exit();
+    }
+
     $sql = "delete from checklist where ch_id='" . $_GET['id'] . "'";
     mysqli_query($connect, $sql);
     echo "<script>alert('ลบข้อมูลเรียบร้อย');</script>";
