+++ customer_save.php
@@ -23,6 +23,21 @@ if($_POST['mode']=="add"){
     mysqli_query($connect, $sql);
     echo "<script>alert('แก้ไขข้อมูลเรียบร้อย');</script>";
 }elseif($_GET['mode']=="delete"){
+
+    $sql_id1 = "select * from reservation where cus_id='".$_GET['id']."'";
+    $query_id1 = mysqli_query($connect, $sql_id1);
+    $num1 = mysqli_num_rows($query_id1);
+
+    $sql_id2 = "select * from get_car where cus_id='".$_GET['id']."'";
+    $query_id2 = mysqli_query($connect, $sql_id2);
+    $num2 = mysqli_num_rows($query_id2);
+
+    if($num1 >=1 or $num2 >=1){
+        echo "<script>alert('ไม่สามารถลบข้อมูลได้ เนื่องจากข้อมูลนี่ถูกใช้แล้ว');</script>";
+        echo "<META http-equiv='refresh' Content='0; URL=customer_list.php'> ";
+        exit();
+    }
+
     $sql = "delete from customer where cus_id='" . $_GET['id'] . "'";
     mysqli_query($connect, $sql);
     echo "<script>alert('ลบข้อมูลเรียบร้อย');</script>";
