+++ employee_save.php
@@ -27,6 +27,20 @@ if($_POST['mode']=="add"){
     $query1= mysqli_query($connect, $sql1);
     $array = mysqli_fetch_array($query1);
     if($array['emp_position']='1'){
+        $sql_id1 = "select * from purchase_order where emp_id='".$_GET['id']."'";
+        $query_id1 = mysqli_query($connect, $sql_id1);
+        $num1 = mysqli_num_rows($query_id1);
+
+        $sql_id2 = "select * from repair where emp_id='".$_GET['id']."'";
+        $query_id2 = mysqli_query($connect, $sql_id2);
+        $num2 = mysqli_num_rows($query_id2);
+
+        if($num1 >=1 or $num2 >=1){
+            echo "<script>alert('ไม่สามารถลบข้อมูลได้ เนื่องจากข้อมูลนี่ถูกใช้แล้ว');</script>";
+            echo "<META http-equiv='refresh' Content='0; URL=employee_list.php'> ";
+            exit();
+        }
+
         $sql = "delete from employee where emp_id='" . $_GET['id'] . "'";
         mysqli_query($connect, $sql);
     }else{
@@ -38,6 +52,21 @@ if($_POST['mode']=="add"){
             echo "<meta http-equiv='refresh' content='0;URL=employee_list.php'>";
             exit();
         }
+
+        $sql_id1 = "select * from purchase_order where emp_id='".$_GET['id']."'";
+        $query_id1 = mysqli_query($connect, $sql_id1);
+        $num1 = mysqli_num_rows($query_id1);
+
+        $sql_id2 = "select * from repair where emp_id='".$_GET['id']."'";
+        $query_id2 = mysqli_query($connect, $sql_id2);
+        $num2 = mysqli_num_rows($query_id2);
+
+        if($num1 >=1 or $num2 >=1){
+            echo "<script>alert('ไม่สามารถลบข้อมูลได้ เนื่องจากข้อมูลนี่ถูกใช้แล้ว');</script>";
+            echo "<META http-equiv='refresh' Content='0; URL=employee_list.php'> ";
+            exit();
+        }
+
         $sql = "delete from employee where emp_id='" . $_GET['id'] . "'";
         mysqli_query($connect, $sql);
         echo "<script>alert('ลบข้อมูลเรียบร้อย');</script>";
